feat: #918 all partnerunit attributes as readable
