+++ src/a06_belief_logic/test/_util/test_a06_str.py
@@ -1,4 +1,5 @@
 from src.a06_belief_logic.test._util.a06_str import (
+    _groupunits_str,
     _keeps_buildable_str,
     _keeps_justified_str,
     _offtrack_fund_str,
@@ -40,6 +41,7 @@ from src.a06_belief_logic.test._util.a06_str import (
 
 def test_str_functions_ReturnsObj():
     # ESTABLISH / WHEN / THEN
+    assert _groupunits_str() == "_groupunits"
     assert _keeps_buildable_str() == "_keeps_buildable"
     assert _keeps_justified_str() == "_keeps_justified"
     assert _offtrack_fund_str() == "_offtrack_fund"
